Restructured tests to run only one build per test file
