+++ .gitignore
@@ -31,4 +31,4 @@
 .DS_Store
 Thumbs.db
 /test/fixtures/*/dist
-/test/fixtures/watch
+/test/fixtures/watch-*
