+++ .vscode/settings.json
@@ -9,6 +9,6 @@
 		"node_modules": true,
 		"package-lock.json": true,
 		"test/fixtures/*/dist": true,
-		"test/fixtures/watch": true
+		"test/fixtures/watch-*": true
 	}
 }
