+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@wildpeaks/glob-webpack-plugin",
-  "version": "1.1.0",
+  "version": "2.0.0",
   "description": "Webpack config to define entry with a glob",
   "author": "Cecile Muller",
   "license": "MIT",
