+++ test/fixtures/watch/src/initial-1.js
@@ -1,0 +1,2 @@
+'use strict';
+console.log('INITIAL 1');
