+++ test/fixtures/watch/src/initial-2.js
@@ -1,0 +1,2 @@
+'use strict';
+console.log('INITIAL 2');
